Add user test routes using dev api route
